requirements.txt
pybit
pandas
ta
python-dotenv
colorama


